+++ apps/roam/src/data/defaultDiscourseRelations.ts
@@ -35,6 +35,16 @@ const DEFAULT_RELATION_VALUES: InputTextNode[] = [
                   { text: "is a", children: [{ text: "destination" }] },
                 ],
               },
+              {
+                text: "Node Positions",
+                children: [
+                  { text: "0", children: [{ text: "100 57" }] },
+                  { text: "1", children: [{ text: "100 208" }] },
+                  { text: "2", children: [{ text: "100 345" }] },
+                  { text: "source", children: [{ text: "281 57" }] },
+                  { text: "destination", children: [{ text: "281 345" }] },
+                ],
+              },
             ],
           },
         ],
@@ -125,6 +135,20 @@ const DEFAULT_RELATION_VALUES: InputTextNode[] = [
                   },
                 ],
               },
+              {
+                text: "Node Positions",
+                children: [
+                  { text: "0", children: [{ text: "250 325" }] },
+                  { text: "1", children: [{ text: "100 325" }] },
+                  { text: "2", children: [{ text: "100 200" }] },
+                  { text: "3", children: [{ text: "250 200" }] },
+                  { text: "4", children: [{ text: "400 200" }] },
+                  { text: "5", children: [{ text: "100 75" }] },
+                  { text: "6", children: [{ text: "250 75" }] },
+                  { text: "source", children: [{ text: "400 325" }] },
+                  { text: "destination", children: [{ text: "400 75" }] },
+                ],
+              },
             ],
           },
         ],
@@ -213,6 +237,20 @@ const DEFAULT_RELATION_VALUES: InputTextNode[] = [
                     text: "is a",
                     children: [{ text: "destination", children: [] }],
                   },
+                ],
+              },
+              {
+                text: "Node Positions",
+                children: [
+                  { text: "0", children: [{ text: "250 325" }] },
+                  { text: "1", children: [{ text: "100 325" }] },
+                  { text: "2", children: [{ text: "100 200" }] },
+                  { text: "3", children: [{ text: "250 200" }] },
+                  { text: "4", children: [{ text: "400 200" }] },
+                  { text: "5", children: [{ text: "100 75" }] },
+                  { text: "6", children: [{ text: "250 75" }] },
+                  { text: "source", children: [{ text: "400 325" }] },
+                  { text: "destination", children: [{ text: "400 75" }] },
                 ],
               },
             ],
