+++ apps/roam/src/index.ts
@@ -1,45 +1,29 @@
-import {
-  addStyle,
-  createHTMLObserver,
-  createButtonObserver,
-  getPageTitleValueByHtmlElement,
-} from "roamjs-components/dom";
-import { createBlock } from "roamjs-components/writes";
+import { addStyle } from "roamjs-components/dom";
 import { render as renderToast } from "roamjs-components/components/Toast";
 
 import { runExtension } from "roamjs-components/util";
 
-import { renderTldrawCanvas } from "./components/canvas/Tldraw";
-import { renderQueryPage, renderQueryBlock } from "./components/QueryBuilder";
 import { QueryBuilderLoadedToast } from "./components/toastMessages";
 
 import runQuery from "./utils/runQuery";
 import isDiscourseNode from "./utils/isDiscourseNode";
 import { fireQuerySync } from "./utils/fireQuery";
 import parseQuery from "./utils/parseQuery";
+import refreshConfigTree from "./utils/refreshConfigTree";
 
-import initializeDiscourseGraphsMode from "./discourseGraphsMode";
 import styles from "./styles/styles.css";
 import settingsStyles from "./styles/settingsStyles.css";
+import discourseGraphStyles from "./styles/discourseGraphStyles.css";
 
 import { registerCommandPaletteCommands } from "./settings/commandPalette";
 import { createSettingsPanel } from "./settings/settingsPanel";
 
-import { renderNodeConfigPage } from "./settings/configPages";
-import { isCurrentPageCanvas as isCanvasPage } from "./utils/isCanvasPage";
-import { isDiscourseNodeConfigPage as isNodeConfigPage } from "./utils/isDiscourseNodeConfigPage";
-import { isQueryPage } from "./utils/isQueryPage";
 import { listActiveQueries } from "./utils/listActiveQueries";
 import { registerSmartBlock } from "./utils/registerSmartBlock";
-import isFlagEnabled from "./utils/isFlagEnabled";
-
-import {
-  enablePageRefObserver,
-  addPageRefObserver,
-  getPageRefObserversSize,
-  previewPageRefHandler,
-  overlayPageRefHandler,
-} from "./utils/pageRefObserverHandlers";
+import { initObservers } from "./utils/initializeObserversAndListeners";
+import { addGraphViewNodeStyling } from "./utils/graphViewNodeStyling";
+import { setQueryPages } from "./utils/setQueryPages";
+import initializeDiscourseNodes from "./utils/initializeDiscourseNodes";
 
 export const DEFAULT_CANVAS_PAGE_FORMAT = "Canvas/*";
 
@@ -62,63 +46,20 @@ export default runExtension(async (onloadArgs) => {
       timeout: 500,
     });
   }
+
   const { extensionAPI } = onloadArgs;
+
   const style = addStyle(styles);
+  const dgraphStyles = addStyle(discourseGraphStyles);
   const settingsStyle = addStyle(settingsStyles);
 
-  const cleanupDiscourseGraphs =
-    await initializeDiscourseGraphsMode(onloadArgs);
-
-  // Observers and Listeners
-  const pageTitleObserver = createHTMLObserver({
-    tag: "H1",
-    className: "rm-title-display",
-    callback: (e) => {
-      const h1 = e as HTMLHeadingElement;
-      const title = getPageTitleValueByHtmlElement(h1);
-      const props = { title, h1, onloadArgs };
-
-      if (isNodeConfigPage(title)) renderNodeConfigPage(props);
-      else if (isQueryPage(props)) renderQueryPage(props);
-      else if (isCanvasPage(props)) renderTldrawCanvas(props);
-    },
-  });
-
-  const queryBlockObserver = createButtonObserver({
-    attribute: "query-block",
-    render: (b) => renderQueryBlock(b, onloadArgs),
-  });
-
-  const pageActionListener = ((
-    e: CustomEvent<{
-      action: string;
-      uid: string;
-      val: string;
-      onRefresh: () => void;
-    }>,
-  ) => {
-    if (!/page/i.test(e.detail.action)) return;
-    window.roamAlphaAPI.ui.mainWindow
-      .getOpenPageOrBlockUid()
-      .then((u) => u || window.roamAlphaAPI.util.dateToPageUid(new Date()))
-      .then((parentUid) => {
-        createBlock({
-          parentUid,
-          order: Number.MAX_VALUE,
-          node: { text: `[[${e.detail.val}]]` },
-        });
-      });
-  }) as EventListener;
+  const { observers, listeners } = await initObservers({ onloadArgs });
+  const [pageActionListener, hashChangeListener, nodeMenuTriggerListener] =
+    listeners;
   document.addEventListener("roamjs:query-builder:action", pageActionListener);
+  window.addEventListener("hashchange", hashChangeListener);
+  document.addEventListener("keydown", nodeMenuTriggerListener);
 
-  if (isFlagEnabled("preview")) addPageRefObserver(previewPageRefHandler);
-  if (isFlagEnabled("grammar.overlay")) {
-    addPageRefObserver((s) => overlayPageRefHandler(s, onloadArgs));
-  }
-  if (!!getPageRefObserversSize()) enablePageRefObserver();
-
-  // Window
-  // @ts-ignore
   window.roamjs.extension.queryBuilder = {
     runQuery: (parentUid: string) =>
       runQuery({ parentUid, extensionAPI }).then(
@@ -132,22 +73,30 @@ export default runExtension(async (onloadArgs) => {
     isDiscourseNode: isDiscourseNode,
   };
 
+  initializeDiscourseNodes();
+  refreshConfigTree();
+  addGraphViewNodeStyling();
   registerCommandPaletteCommands(onloadArgs);
   createSettingsPanel(onloadArgs);
-  registerSmartBlock(extensionAPI);
+  registerSmartBlock(onloadArgs);
+  setQueryPages(onloadArgs);
 
   return {
-    elements: [style, settingsStyle],
-    observers: [pageTitleObserver, queryBlockObserver],
+    elements: [style, settingsStyle, dgraphStyles],
+    observers: observers,
     unload: () => {
       window.roamjs.extension?.smartblocks?.unregisterCommand("QUERYBUILDER");
-      cleanupDiscourseGraphs();
       // @ts-ignore - tldraw throws a warning on multiple loads
       delete window[Symbol.for("__signia__")];
       document.removeEventListener(
         "roamjs:query-builder:action",
         pageActionListener,
       );
+      window.removeEventListener("hashchange", hashChangeListener);
+      document.removeEventListener("keydown", nodeMenuTriggerListener);
+      window.roamAlphaAPI.ui.graphView.wholeGraph.removeCallback({
+        label: "discourse-node-styling",
+      });
     },
   };
 });
